Moved GlooCache to new file; Functions renames; Minor fixes
