+++ torch/lib/THD/base/DataChannel.hpp
@@ -12,23 +12,8 @@
 #include <utility>
 
 
-namespace std {
-
-template<>
-struct hash<THDReduceOp> {
-  std::size_t operator()(const THDReduceOp& op) const {
-    return hash<int>()(static_cast<int>(op));
-  }
-};
-
-template<>
-struct hash<thpp::Type> {
-  std::size_t operator()(const thpp::Type& type) const {
-    return hash<char>()(static_cast<char>(type));
-  }
-};
-
-} // namespace std
+MAKE_HASHABLE(THDReduceOp, static_cast<int>(t));
+MAKE_HASHABLE(thpp::Type, static_cast<char>(t));
 
 
 namespace thd {
